+++ packages/types-support/src/metadata/static-kusama.ts
@@ -1,6 +1,6 @@
 // Copyright 2017-2021 @polkadot/types-support authors & contributors
 // SPDX-License-Identifier: Apache-2.0
 
-import metadata from './v13/polkadot-hex';
+import metadata from './v13/kusama-hex';
 
 export default metadata;