+++ CHANGELOG.md
@@ -5,8 +5,9 @@
 Changes:
 
 - Handle `BitSequence` in `PortableRegistry`
-- Move all test-support static files into `@polkadot/types-support`
-- Align static Metadata with latest Substrate
+- Move all static test support files into `@polkadot/types-support`
+- Align static Metadata with latest Substrate, Polkadot & Kusama
+- Adjust v14 Metadata definitions to latest version
 
 
 ## 5.3.2 Aug 3, 2021